requirement.txt
transformers
streamlit
langchain-community
tensorflow
python-dotenv
dataset
sentencepiece
huggingface
matplotlib